requirements.txt
social-auth-app-django
django-ckeditor
pillow
pytils
python-dateutil